move mock seata server to fixture package
