+++ sharding-transaction/sharding-transaction-base/sharding-transaction-base-seata-at/src/test/java/org/apache/shardingsphere/transaction/base/seata/at/fixture/MockSeataServer.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.transaction.base.seata.at;
+package org.apache.shardingsphere.transaction.base.seata.at.fixture;
 
 import io.netty.bootstrap.ServerBootstrap;
 import io.netty.buffer.PooledByteBufAllocator;
